benchmark/pyproject.toml
[project]
name = "ocr-benchmark"
version = "1.0.0"
description = "OCR benchmark for PaddleOCR models"
requires-python = ">=3.11,<3.13"

dependencies = [
    # PaddlePaddle 3.x for macOS ARM
    "paddlepaddle>=3.0.0",
    # PaddleOCR 3.x with mobile/server model support
    "paddleocr>=3.0.0",
    # Image processing
    "Pillow>=10.0.0",
    "numpy",
    "pyspellchecker>=0.8.4",
]

[project.scripts]
ocr-benchmark = "ocr_benchmark:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
